// api/types.gen.ts
export type JsonSchema = {
  /**
   * Json Schema
   * A JSON Schema object. Only a subset of JSON Schema is supported.
   */
  json_schema: {
    [key: string]: unknown
  }
  /**
   * Type
   * The type of schema being defined. Always `json`.
   */
  type?: 'json'
}

/**
 * TaskRun
 * Status of a task.
 */
export type TaskRun = {
  /**
   * Run Id
   * ID of the task run.
   */
  run_id: string
  /**
   * Status
   * Status of the run.
   */
  status:
    | 'queued'
    | 'action_required'
    | 'running'
    | 'completed'
    | 'failed'
    | 'cancelling'
    | 'cancelled'
  /**
   * Is Active
   * Whether the run is currently active; i.e. status is one of {'running', 'queued', 'cancelling'}.
   */
  is_active: boolean
  /**
   * Warnings
   * Warnings for the run.
   */
  warnings?: Array<ShapleySearchTasksApiV1Warning> | null
  /**
   * Processor
   * Processor used for the run.
   */
  processor: string
  /**
   * Metadata
   * User-provided metadata stored with the run.
   */
  metadata?: {
    [key: string]: string | number | number | boolean
  } | null
  /**
   * Created At
   * Timestamp of the creation of the task, as an RFC 3339 string.
   */
  created_at: string | null
  /**
   * Modified At
   * Timestamp of the last modification to the task, as an RFC 3339 string.
   */
  modified_at: string | null
}
// ...
export type ShapleySearchTasksApiV1Warning = {
  /**
   * Type
   * Type of warning. Note that adding new warning types is considered a backward-compatible change.
   */
  type: string
  /**
   * Message
   * Human-readable message.
   */
  message: string
  /**
   * Detail
   * Optional detail supporting the warning.
   */
  detail?: {
    [key: string]: unknown
  } | null
}